a lot of stuff done
messed with obj_controller, but ultimately didn't change anything with that.   Rooms are going under constant renovation, and I messed with the GUI a bit.   The player's hud has been temporarily removed, as I am trying to find what would be best for the style of gameplay.   The cursor over the enemy's head has been tweaked a bit;  before, when you would release the kill button, the arrow would turn white for a split second and disappear.  That is no longer the case, and it disappears as soon as the enemy is dead